Better italicisation
Both within ghini and for reports etc.

The 2 functions had originally been used in a completely unrelated
project but with a little tweaking seem to fit the bill here.  Could
possibly improve.  The main use is for when we need to add anything
other than a basic species epithet in the species field.

It could be improved.  E.g. diaeresis (ë etc.) is allowed in the ICBN
but I haven't accounted for them as I don't have any names that use
them.