More of 3
Note: Some of this is just cosmetic or from linters.
- PrefsView was failing at updating the view. It seems that you need to
use GObject.idle_add() to schedule GUI task to executed in the main
thread, see:https://stackoverflow.com/a/39558651 for instance.
- I realised I need to update the view when the config file is updated.
- AccelGroup object: connect_groupis now connect (this is just for
keyboard shortcuts to menu items, didn't test much but com-Q, F1 seems
to work now and I'm not getting the error.