bits and pieces on 3
pep8/linter stuff.
get connmgr working for version checks.
more refinements on FileChooserNative:
	- in windows its best to set the current_folder to something
	  like home or the template root etc..
	- changed the last few filechoosers to native.
reports: when starting a new report (or if settings are not yet set)
make sure all the fields are cleared, makes sure we get a sane start
point for filechooser.
template_downloader:
	- switch to using pathlib.Path over os.path, avoids mixing
	  posix/windows path separators.
	- allow the user to bailing out.
	- clean up non template files when finished unzipping.
utils.clear_model - this was causing a large bottleneck and requiring
many calls to the DB on every run.  It was put in place to avoid a
memory leak but I can't find evidence that this is still an issue so now
it just deletes the model and sets to None.