debug logging as a pref, also more 'as' imports
Something I had half started a while ago.  Allow the user to set the
logging to debug via a preference setting at module level.  Handy for
when working with a strict environment with a frozen build and no way to
access the code.

Also, a few more changing 'as' imports to 'from' imports.