Accession editor, if no species avoid committing
In a new database if you select to insert an accession the editor warns
the user that they must add taxonomy first then only allows OK to close
the window.  This was leaving an new Accession() still in the session,
after closing the window with OK the __del__ method of
GenericModelViewPresenterEditor would try to commit it causing errors.