postgres fix and geography cast to str
After I had a accidentally introduced a bug to the code then opened a
postgresql db plugins where dropped and db was left in a state with a
dropped geography table causing it to fail installing the plugins again.
This attempts to allow a recovery from that state.
Also, realised that I had not cast int tdwg codes to str (cast all for
good measure in future) and hadn't closed the session in geography
imports causing a dead lock.