Avoid multiple instances of the _prefs class
Have come across several issues with tests where the result depends on
the settings in the system config file (e.g. in the last commit if
return_accepted was set False some tests would fail.) or where running
tests changed the users settings.

By creating a temp config file for tests and making sure not to import
the pref variable directly (which will cause another instance to be
created with the system config files) this is avoided.  Treating it as a
kind of singleton.