Cultivar in binomial search + warn user dangerous
BinomialNameAction will recognise a cultivar name in the second part if
it begins with a single ' or is the complete name (i.e. "Gr 'Al" or "Gr
'Alex Pink'" will both find "Grevillea 'Alex Pink' with no species
epithet entry").

When ValueListAction receives a list of tokens that contain a single
letter it can taken an inordinate amount of (blocking) time to find
results.  At least allow the user to bail if they didn't intend to do
such a random search.

also:
QueryBuilder don't quote plain % _ and treat spinbuttons as entries
doesn't mangle None searches etc. when not changed.