Completions improvements.
More along the line of commit 0d7db090. This time looking more at how
species editor and genus field.

Some simplification as seems idle_add is neither needed or desirable for
responsiveness (i.e. when typing a genus and we have a match it should
immediately be apparent by displaying in the fullname label).

Similarly, if the user deletes a letter after it is matched it should
immediately be removed (on_select(None)).