prefs - none existent sections, refresh history
When prefs are edited they may change the history entries in the main
entry (especially if restoring a previous config file) so repopulate.
If itersection is called on a none existent section it will fail. (i.e.
if web_button_defs.family section is deleted then the family infobox
would fail trying to iterate the none existent section.)