Fix Gtk error switching SearchView/previous view.
Switching between SearchView and a previous view and back could
produce:
Gtk-CRITICAL **: _gtk_accel_group_attach: assertion 'g_slist_find
(accel_group->priv->acceleratables, object) == NULL' failed

Due to SearchView.accel_group.  This was no longer in use so have
removed it.

Error was due to the fact that in bauble.command_handler the accel_group
is removed/added correctly but this was not happening when setting
previous view in bauble.ui.GUI.set_view.  (If ever reimplemented would
need to address this.)