Species add label_markup field
An optional field similar to label_distribution that can be used to
override the default markup for a species.  Intended for use in reports
etc. where the name may need to be abbreviated etc.
e.g.:
	Brachychiton sp. (Ormeau L.H.Bird AQ435851)
	Brachychiton sp. (Ormeau)
	or
	Cynodon dactylon × transvaalensis 'DT-1' (PBR) TIFTUF™
	Cynodon TIFTUF™ (PBR)
Only accepts valid markup (as accepted by Pango due to the label that
displays the result).  Won't save the same value as the default markup,
will display the text in blue to signify this.
+ adjusted some reports (+ removed white space in one)

+ decided to move away from using Pango specific xx-small for PBR
  instead going for just <small> (and weight=light in search view)