Fix fresh install failing to save connection
When no pictures/documents_path values are set in prefs file meta is
queried.  As db.Session is still None when connection manager is
launched it would fail (in check_create_paths) and hence not save the
preferences at all.

Changes here allow the defaults to be returned in this situation so this
does not fail.

(NOTE if a global entry has been set and hence the pref entry is '' for
root check_create_paths does not check the sub directories anyway -this
is the users/administrators responsibility to check that a global
directory exists.  Also, the only time, at this stage of startup, that
it would matter if there was a prefs entry different to the default is
if a user has set it in their prefs file, which would not fail anyway.)