Appveyor set force_update true
force_update updates the release details (contrary to the documentation
stating it is to 'overwrite files in an existing release' release
artifacts are always uploaded and this setting refers to the release
details. i.e. without it set to true you do not get the line '10 commits
to XYZ since this release'.)