Species editor, make syn history entry last
If the species name is changed and user selects to keep the previous
name as a synonym the history entry for inserting the copy of the old
species would come first.  Does not matter in at the time but If this
was done in a clone and then the clone synced this it would create a
unique constraint error on the full_sci_name because the original
species still exists (NOTE sync commits one history line at a time to
pick up conflicts and offer the user a chance to resolve them).