Search get aggregate functions to work as expected
in more complex queries...
After discovering this search string would fail:

species where distribution.geography.name = 'Queensland' and
count(distribution.geography.id) > 1

I realised the group by and having portions would be skipped in
SearchAndAction (not a where or except clause) unless the first clause
was the count() portion.  Even if it was first the group by portion
could change the outcome of the other clauses as group by is executed
before select.  Converting to using a where with subquery approach
gives
results more as would be expected (although its most likely not as
efficient).