Allow exporting an accessions qualified name
Provide a hybrid_property that returns species_str() i.e. species string
with ID qualifier in place.  As this hybrid_property does not have an
expression or setter it can not be searched or set (e.g. via imports) so
don't make it or similar hybrid_properties available in these contexts.

+ a little tidying up