HistoryView lazy load rows,
Moving from ordering by timestamp (which had been the case since
inception of HistoryView) to ordering by `id` (see bc0edb9c) massively
increased the speed of retrieval but also tended to freeze the window on
large history tables (e.g. 100 000 lines) as multiple queries were sent
to the database in rapid succession.
This commit instead lazy loads columns as needed when the user scrolls
and avoids so many database hits and the need for running in a thread.