Add location search map to splash screen
Is only added if institution has lats/longs.
Locations are labelled with their code for visual reference.
Generates a search for the location(s) clicked on.

Added a simplified ray_casting `is_point_within_poly` function to
utils.geo, could be improved but this version seems easier to grok than
some and serves the purpose.

+ some extra type annotations (or corrections), etc.