Fix pointless history entries
Caused by loading deferred geojson column without a change.  Was causing
single list changes.  Also, when changing the value if the column is not
loaded first then history would record a single list change with no
indication of its prior state.  This commit make sure to touch the
record if the change is occurring (PresenterMapMixin) and avoid touching
it when not (DistributionMap, SearchViewMapPresenter, LocationSearchMap)

+ upgrade sqlalchemy.  Was receiving strange error, only when using
  coverage, in an unrelated test:
  `test_plants.py::GeographyTests::test_get_species - KeyError:
  "Deferred loader for attribute '_created' failed to populate
  correctly"`
  Upgrading appears to fix this.