Garden map - allow manual refresh
Intended use case is if user adds an item to the search view root (via
clicking on an infobox clickable item adds that uses
select_in_search_results) then refreshing will add its plants to the map
also. (i.e. search a plant, click its location in the infobox then
refresh to see all plants in that location on the map.)  This is not a
default behaviour because in most cases the user wants to see the
details without polluting the map.

Also, one extra _update_worker test to complete coverage.