Search allow filter by `in`, `like`, `contains`
Not using `lower()` in `OPERATIONS.get()` was not finding the correct
operator function.  Have documented OPERATIONS to avoid this in future.
Also, realised there was no reason not to allow `in` in these filters
(NOTE if using comma separation you need to put the in clause last when
multiple filter clauses or else don't use comma separation)

Also:
Realised that InSetClause was not needed, BinaryClause works fine in its
place.
Extra test for a recursive filtered query that I find useful and do not
want to break in the future.
BNF corrections.