_hide_restore_pic_pane restore once
Don't use timer on size-allocate, instead bail unless its the first run
after _hide_restore_pic_pane and the last result was successful.

aa26844b introduced a glitch making it almost impossible to change the
pane size as it was size-allocate is firing while the user was trying
to change the slider position.