PlantEditor allow save without closing the editor.
Also, add label to display ID of the plant.
Intention is: when working with a shapefile map (QGIS, etc.) the user
can create a plant to match a point on on the map then copy and add the
ID to the shapefile to ease importing the spatial data from the
shapefile without leaving the editor.