SearchView don't create, cache infobox instances
SearchView should not need to know about infoboxes other which one to
use and then calling update on them.  Here the plugin is responsible for
instantiating the correct one for each type.  This is done once for the
life of the app to avoid garbage collection issues (i.e. having to
destroy on a database connection change)

Also:
- Don't check for GardenPlugin and remove widgets, can assume it is a
  core plugin, and removing the widget only has potential to cause
  further issues.
- Reordered some lines so it is easier to follow. (button_defs,
  PropertiesExpanded after InfoBoxPage)
- Should be able to use a single custom column in
  GeneralSpeciesExpander, previous behaviour was both or none.
- Remove SpeciesInfoPage, was not using anyway, include test for
  InfoBoxes that ARE tabbed and InfoBoxPage, although not currently
  using tabs I can see one day they may.
- Extra test, link button with fields (was previously being covered but
  not after changes here? SpeciesInfoPage?)
- More type annotations.