Refactoring tag plugin pt5 (+ SearchView)
Move tags bottom page to Gtk.Template.

To accomplish this the same was done for notes bottom page and
associated portions of SearchView also refactored.

SearchView.bottom_info was using ViewMeta as little more than a dict,
now removed along with the methods that relied on it.
In its place have used an approach where SearchView only needs to be
supplied a label for the tab and the notebook page widget instance (no
need for for a row-activated handler, sorter, tree, fields used, etc.).
Now SearchView only needs to call update on the widget when appropriate
and leave the widget itself to deal with the rest (increasing cohesion
and reducing coupling).