Make sort by taxon searches faster
Got to love it when one simple change makes an order of magnitude
difference! (A search for all accessions, sorted by taxon, could take
over 6 seconds, now around half a second.)
When sorting accessions or plants by the species string the database was
queried for each item.  This loads the species with the initial query so
there is no need to go back out to the database for them.