SearchView replace mulitproc_counter
Taking a new approach to top_level_count.
The approach here is to query the database directly once per search.
Also, puts the logic in the domain classes and SearchView only needs
to know domain class to call (reduced coupling, increased cohesion) for
the progress bar string.
Also, provides a uniform output for all domains except Geography and
Tags (which can not be queried in a way that represents SearchView so
easily) and provides the result as a dataclass which can be used
elsewhere (e.g. reports).

+ removed count fast pref option, should no longer be needed.