GenericPresenter - problems-changed signal
Move emitting the problems-changed custom signal into the base class to
make it easier to use in more places and more consistent.
For convenience a default definition for __gsignals__ is provided which
can be included in the subclass if choosing to use it.  If no definition
for problems-changed is supplied nothing is emitted.