GeographyMenu improvements
Relates to 7902d155

Editors that use a GeographyMenu (species, accession), on a slow system,
could take and inordinate amount of time to start.

This version attempts to query the database and organise the geographies
only once per session and, by using a separate thread, does not block
the editor from displaying (although the button the menu is attached to
may remain insensitive for a period).

If the geography table is updated at any point (should be rare) the menu
is reset for the next time it is used.