DistributionMap - only detach image from cache
Only if the image comes from the cache could it already be attached to a
parent widget (e.g. when a species with a single geography distribution
and the geography itself are both in the search view). So only need to
remove it from the parent if it comes from the cache. See 5307b25a