NoteBox spell checking, get Gspell to work, mingw.
Ensure Enchant providers are available where expected, include the
enchant.ordering file (needs runtime hook) and collect the aspell
dictionaries.

For this to work, after pyinstaller has finished we need to brute force
move lib/enchant-2 to ../lib/enchant-2 outside of _internal as
libenchant expects to find the providers at ../lib/enchant relative to
itself.  Also means we need to account for the extra directory with the
NSIS uninstaller.  Hope to find a better solution.