GenericPresenter refactor to allow more complex
Use non-data descriptors to define methods for assembling UI components
as needed, handling most of the common actions (logging, problems,
validating, exception handling, notifying the user, etc.) in one place
and avoiding repeated code etc..  Sane defaults are provided for simple
use cases.

This is not settled yet, consider it a concept at this point.
Considering the inheritance model used I realise this could lead to a
proliferation of classes.  With the limited set of widgets it could be
applied to I think it unlikely.

Pylint does not handle method descriptors requiring disabling
not-callable.

Fix garbage collection issues here and elsewhere by calling `collect`
explicitly when a dialog is destroyed.